feat(Data): expandir ClassDataTable com FeatureType e SubclassNames
Adiciona campo FeatureType em FClassFeature para diferenciar features
automáticas de escolhas do jogador (Automatic, SubclassSelection, ASI, Choice).

Adiciona campo SubclassNames em FClassDataRow para suportar subclasses
por classe (ex: Champion, Battle Master para Fighter).

Alinha com arquitetura Data-Driven e mantém compatibilidade retroativa.

+++ Source/MyProject2/Characters/Components/CharacterDataComponent.h
@@ -42,7 +42,7 @@ public:
 
     /** Nível total do personagem */
     UPROPERTY(BlueprintReadOnly, Category = "Character Data")
-    int16 CharacterTotalLvl = 0;
+    int32 CharacterTotalLvl = 0;
 
     /**
      * Faz log da ficha do personagem no console.
